added functionality to end gateway
