+++ src/main/resources/assets/gilly7ce-carpet-addons/lang/en_us.json
@@ -10,5 +10,5 @@
   "carpet.rule.netheritePickaxeInstantMineEndStone.desc": "A netherite pickaxe with efficiency V combined with the haste II status effect will instant mine end stone type blocks.",
   "carpet.rule.netheritePickaxeInstantMineNetherBricks.desc": "A netherite pickaxe with efficiency V combined with the haste II status effect will instant mine nether brick type blocks.",
   "carpet.rule.phantomsObeyHostileMobCap.desc": "Phantoms will no longer spawn if the hostile mobcap is full. This is per player.",
-  "carpet.rule.spectatorPortalFix.desc": "Spectator will be able to go through portals."
+  "carpet.rule.spectatorPlayersUsePortals.desc": "Spectator players can go through nether portals, end portals and end gateways."
 }
