+++ src/main/resources/gilly7ce-carpet-addons.mixins.json
@@ -9,7 +9,7 @@
     "PistonBlock_MovableEmptyEndPortalFrameMixin",
     "ServerChunkManager_PhantomsObeyHostileMobCapMixin",
     "ServerPlayerEntity_InstantMineMixin",
-    "SpectatorPortalFix",
+    "ServerPlayerEntity_SpectatorPlayersUsePortalsMixin",
     "invokers.SpawnHelperInfoInvokerMixin"
   ],
   "injectors": {
